+++ sync-for-payables/CodatSyncPayables/Models/Operations/DeleteBillPaymentRequest.cs
@@ -26,5 +26,11 @@ namespace CodatSyncPayables.Models.Operations
         /// </summary>
         [SpeakeasyMetadata("pathParam:style=simple,explode=false,name=companyId")]
         public string CompanyId { get; set; } = default!;
+
+        /// <summary>
+        /// Unique identifier for a connection.
+        /// </summary>
+        [SpeakeasyMetadata("pathParam:style=simple,explode=false,name=connectionId")]
+        public string ConnectionId { get; set; } = default!;
     }
 }