+++ sync-for-payables/CodatSyncPayables/Models/Shared/BillAccountingPaymentAllocationAllocation.cs
@@ -12,7 +12,7 @@ namespace CodatSyncPayables.Models.Shared
 {
     using Newtonsoft.Json;
     
-    public class BillPaymentAllocationAllocation
+    public class BillAccountingPaymentAllocationAllocation
     {
 
         /// <summary>