+++ sync-for-payables/RELEASES.md
@@ -77,3 +77,13 @@ Based on:
 - [csharp v2.2.0] sync-for-payables
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Sync.Payables/2.2.0 - sync-for-payables
+
+## 2023-11-06 01:28:37
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.3.0] sync-for-payables
+### Releases
+- [NuGet v2.3.0] https://www.nuget.org/packages/Codat.Sync.Payables/2.3.0 - sync-for-payables